Fixes for Windows x64 builds. Use int instead of size_t as field width specifier in printf. Include stdio.h before CUnit.h in Win.c to prevent a complaint about snprintf being defined as a macro.

+++ CUnit/Sources/Console/Console.c
@@ -476,7 +476,7 @@ static void list_suites(CU_pTestRegistry pRegistry)
 {
   CU_pSuite pCurSuite = NULL;
   int i;
-  static size_t width[6];
+  static int width[6];
 
   if (NULL == pRegistry) {
     pRegistry = CU_get_registry();
@@ -492,12 +492,12 @@ static void list_suites(CU_pTestRegistry pRegistry)
 
   /* only need to calculate formatting widths once */
   if (0 == width[0]) {
-    width[0] = CU_number_width(pRegistry->uiNumberOfSuites) + 1;
+    width[0] = (int)CU_number_width(pRegistry->uiNumberOfSuites) + 1;
     width[1] = 34;
-    width[2] = CU_MAX(strlen(_("Init?")), CU_MAX(f_yes_width, f_no_width)) + 1;
-    width[3] = CU_MAX(strlen(_("Cleanup?")), CU_MAX(f_yes_width, f_no_width)) + 1;
-    width[4] = CU_MAX(strlen(_("#Tests")), CU_number_width(pRegistry->uiNumberOfTests) + 1) + 1;
-    width[5] = CU_MAX(strlen(_("Active?")), CU_MAX(f_yes_width, f_no_width)) + 1;
+    width[2] = (int)CU_MAX(strlen(_("Init?")), CU_MAX(f_yes_width, f_no_width)) + 1;
+    width[3] = (int)CU_MAX(strlen(_("Cleanup?")), CU_MAX(f_yes_width, f_no_width)) + 1;
+    width[4] = (int)CU_MAX(strlen(_("#Tests")), CU_number_width(pRegistry->uiNumberOfTests) + 1) + 1;
+    width[5] = (int)CU_MAX(strlen(_("Active?")), CU_MAX(f_yes_width, f_no_width)) + 1;
   }
 
   fprintf(stdout, "\n%s",   _("--------------------- Registered Suites -----------------------------"));
@@ -533,7 +533,7 @@ static void list_tests(CU_pSuite pSuite)
 {
   CU_pTest pCurTest = NULL;
   unsigned int uiCount;
-  static size_t width[3];
+  static int width[3];
 
   assert(NULL != pSuite);
   assert(NULL != pSuite->pName);
@@ -548,10 +548,10 @@ static void list_tests(CU_pSuite pSuite)
   assert(NULL != pSuite->pTest);
 
   /* only number of tests can change between calls */
-  width[0] = CU_number_width(pSuite->uiNumberOfTests) + 1;
+  width[0] = (int)CU_number_width(pSuite->uiNumberOfTests) + 1;
   if (0 == width[1]) {
     width[1] = 34;
-    width[2] = CU_MAX(strlen(_("Active?")), CU_MAX(f_yes_width, f_no_width)) + 1;
+    width[2] = (int)CU_MAX(strlen(_("Active?")), CU_MAX(f_yes_width, f_no_width)) + 1;
   }
 
   fprintf(stdout, "\n%s",
